requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.5.0
pydantic-settings==2.1.0

# NLP and Text Processing
nltk==3.8.1
spacy==3.7.2
textstat==0.7.3
syllables==1.0.7

# LLM Integration
ollama==0.1.6
langchain==0.1.0
langchain-community==0.0.10

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Visualization and Reporting
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
reportlab==4.0.7

# API (Optional)
fastapi==0.108.0
uvicorn==0.25.0
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Development Tools
black==23.12.1
flake8==6.1.0
mypy==1.7.1
jupyter==1.0.0

# Database (Optional)
sqlalchemy==2.0.23
alembic==1.13.1

# Utilities
tqdm==4.66.1
click==8.1.7
rich==13.7.0